@@ -24,7 +24,7 @@
     "url": "https://github.com/Huachao/vscode-restclient.git"
   },
   "engines": {
-    "vscode": "^1.57.0"
+    "vscode": "^1.81.0"
   },
   "categories": [
     "Other",
@@ -674,25 +674,24 @@
   "scripts": {
     "vscode:prepublish": "webpack --mode production",
     "webpack": "webpack --mode development",
-    "watch": "webpack --mode development --watch --info-verbosity verbose",
+    "watch": "webpack --mode development --watch",
     "tslint": "tslint --project tsconfig.json"
   },
   "devDependencies": {
     "@types/aws4": "^1.5.1",
     "@types/fs-extra": "^5.0.4",
-    "@types/got": "^9.6.8",
     "@types/mocha": "^5.2.6",
-    "@types/node": "^12.11.7",
-    "@types/vscode": "^1.57.0",
-    "mocha": "^6.1.4",
-    "ts-loader": "^5.3.2",
+    "@types/node": "^18.0.0",
+    "@types/vscode": "^1.81.0",
+    "mocha": "^10.4.0",
+    "ts-loader": "^7.0.5",
     "tslint": "^6.1.2",
-    "typescript": "^3.7.3",
-    "webpack": "^4.26.0",
-    "webpack-cli": "^3.1.0"
+    "typescript": "^5.4.2",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1"
   },
   "dependencies": {
-    "adal-node": "^0.1.26",
+    "adal-node": "^0.2.4",
     "applicationinsights": "^1.0.5",
     "aws4": "^1.9.1",
     "code-highlight-linenums": "^0.2.1",
@@ -703,7 +702,7 @@
     "encodeurl": "^1.0.1",
     "filesize": "^3.3.0",
     "fs-extra": "^5.0.0",
-    "got": "^9.6.0",
+    "got": "^11.8.6",
     "highlight.js": "^10.4.1",
     "http-proxy-agent": "^2.1.0",
     "https-proxy-agent": "^2.2.3",
@@ -714,10 +713,10 @@
     "mime-types": "^2.1.14",
     "pretty-data": "^0.40.0",
     "tough-cookie": "^4.1.3",
-    "tough-cookie-file-store-bugfix": "Huachao/tough-cookie-file-store-bugfix",
+    "tough-cookie-file-store": "^2.0.3",
     "uuid": "^3.3.2",
     "xmldom": "^0.5.0",
     "xpath": "^0.0.27",
-    "yargs-parser": "~11.1.1"
+    "yargs-parser": "^21.1.1"
   }
 }
