@@ -1,4 +1,4 @@
-import { CancellationToken, Hover, HoverProvider, MarkdownString, MarkedString, Position, TextDocument } from 'vscode';
+import { CancellationToken, Hover, HoverProvider, MarkdownString, Position, TextDocument } from 'vscode';
 import { RequestVariableProvider } from '../utils/httpVariableProviders/requestVariableProvider';
 import { VariableUtility } from '../utils/variableUtility';
 
@@ -14,9 +14,13 @@ export class RequestVariableHoverProvider implements HoverProvider {
 
         const { name, value, warning, error } = await RequestVariableProvider.Instance.get(fullPath, document);
         if (!error && !warning) {
-            const contents: MarkedString[] = [];
+            const contents: MarkdownString[] = [];
             if (value) {
-                contents.push(typeof value === 'string' ? value : { language: 'json', value: JSON.stringify(value, null, 2) });
+                if (typeof value === 'string') {
+                    contents.push(new MarkdownString(value));
+                } else {
+                    contents.push(new MarkdownString().appendCodeblock(JSON.stringify(value, null, 2), 'json'));
+                }
             }
 
             contents.push(new MarkdownString(`*Request Variable* \`${name}\``));
