@@ -144,7 +144,9 @@ export class AadV2TokenProvider {
         const tryAgain = "Try again";
         const done = "Done";
 
-        let value = await window.showInformationMessage(signInPrompt, messageBoxOptions, signIn);
+        type SignInResult = typeof signIn | typeof tryAgain | typeof done | undefined;
+
+        let value: SignInResult = await window.showInformationMessage(signInPrompt, messageBoxOptions, signIn);
         if (value === signIn) {
             do {
                 await this.clipboard.writeText(deviceCodeResponse.user_code);
